# api/openapi.yaml
# Design-time API contract for this service domain — owned BY this repo
# (contract-per-repo, Netflix-style; no central contracts repo).
# The runtime spec served at /v3/api-docs must stay compatible with this file;
# Phase 2 adds contract tests that enforce it.
openapi: 3.0.3
info:
  title: "IFSB Regulatory Reporting — BIAN Semantic API"
  description: >
    BIAN Service Domain 'IFSB Regulatory Reporting' (functional pattern: Process,
    business area: Takaful and Shared Operations) operating on the
    'IFSB Regulatory Report Procedure' control record.
  version: 0.1.0
servers:
  - url: /
    description: direct (in-cluster)
  - url: /isb-ifsb-regulatory-reporting
    description: through the platform gateway (prefix is rewritten away)
tags:
  - name: service-domain
  - name: control-record

paths:
  /v1/service-domain:
    get:
      tags: [service-domain]
      operationId: getServiceDomain
      summary: Service domain metadata
      responses:
        "200":
          description: SD identity and BIAN classification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceDomainInfo"

  /v1/ifsb-regulatory-report-procedure/initiate:
    post:
      tags: [control-record]
      operationId: initiate
      summary: "BIAN action term: Initiate"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Properties"
      responses:
        "201":
          description: Control record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlRecord"

  /v1/ifsb-regulatory-report-procedure:
    get:
      tags: [control-record]
      operationId: list
      summary: "BIAN action term: Retrieve (collection)"
      responses:
        "200":
          description: All control records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ControlRecord"

  /v1/ifsb-regulatory-report-procedure/{crId}/retrieve:
    get:
      tags: [control-record]
      operationId: retrieve
      summary: "BIAN action term: Retrieve"
      parameters:
        - $ref: "#/components/parameters/CrId"
      responses:
        "200":
          description: The control record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlRecord"
        "404":
          description: Unknown control record id

  /v1/ifsb-regulatory-report-procedure/{crId}/update:
    put:
      tags: [control-record]
      operationId: update
      summary: "BIAN action term: Update"
      parameters:
        - $ref: "#/components/parameters/CrId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Properties"
      responses:
        "200":
          description: Updated control record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlRecord"
        "404":
          description: Unknown control record id

  /v1/ifsb-regulatory-report-procedure/{crId}/control:
    put:
      tags: [control-record]
      operationId: control
      summary: "BIAN action term: Control (suspend | resume | terminate)"
      parameters:
        - $ref: "#/components/parameters/CrId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ControlAction"
      responses:
        "200":
          description: Control applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlRecord"
        "400":
          description: Unknown control action
        "404":
          description: Unknown control record id

components:
  parameters:
    CrId:
      name: crId
      in: path
      required: true
      schema:
        type: string
      description: Control record identifier (CR-…)

  schemas:
    ServiceDomainInfo:
      type: object
      properties:
        serviceDomain: { type: string, example: "IFSB Regulatory Reporting" }
        businessArea: { type: string, example: "Takaful and Shared Operations" }
        businessDomain: { type: string, example: "Shared Operations" }
        functionalPattern: { type: string, example: "Process" }
        assetType: { type: string, example: "IFSB Regulatory Report" }
        controlRecord: { type: string, example: "IFSB Regulatory Report Procedure" }
        version: { type: string }
        phase: { type: string }

    ControlRecord:
      type: object
      properties:
        controlRecordId: { type: string, example: "CR-7f3a…" }
        controlRecordType: { type: string, example: "IFSB Regulatory Report Procedure" }
        status:
          type: string
          enum: [INITIATED, ACTIVE, SUSPENDED, TERMINATED]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        properties:
          $ref: "#/components/schemas/Properties"

    Properties:
      type: object
      additionalProperties: true
      description: Free-form domain properties (Phase 2 replaces with the real per-domain schema)

    ControlAction:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [suspend, resume, terminate]